Revised UVA-10642

+++ Lab II/1133317-10642.cpp
@@ -1,10 +1,11 @@
+//Accepted
 #include <iostream>
 #include <vector>
 using namespace std;
 
 int main()
 {
-	int C = 0;
+	long long int C = 0; // modify to long long
 	cin >> C;
 	for (int i = 0; i < C; i++)
 	{
